refactor: AttendancePage 하위 디렉터리 위치 변경
