+++ src/pages/DogDetailGalleryPage/DogGalleryPage.tsx
