+++ src/routes/AdminRoutes.tsx
@@ -35,11 +35,7 @@ const AdminRoutes = ({ queryClient }: { queryClient: QueryClient }): RouteObject
           children: [
             {
               index: true,
-              element: (
-                <Suspense>
-                  <Pages.AttendancePage />
-                </Suspense>
-              )
+              element: <Pages.AttendancePage />
             },
             {
               path: PATH.ADMIN_ATTENDANCE_INFO(),
